changed tests

+++ acp_calc.py
@@ -69,7 +69,7 @@ class AcpBrevet():
 				print("r_dist: {}\nhrs: {}\nmins: {}\nspan: {}\nspeed{}".format(remaining_dist,hrs,mins,span,speed))
 				dist_seg = min(remaining_dist,span)
 				print("dist_seg: {}".format(dist_seg))
-				hrs = hrs + dist_seg / speed
+				hrs = int(hrs + dist_seg / speed)
 				mins = mins + int(round((dist_seg % speed)*MINUTES_PER_HOUR))
 				remaining_dist = remaining_dist - span
 				print("r_dist: {}\nhrs: {}\nmins: {}\nspan: {}\nspeed{}".format(remaining_dist,hrs,mins,span,speed))
